third_party/WebKit/LayoutTests/accessibility/role-change.html

<!DOCTYPE HTML>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<div role="grid">
  <div role="presentation" id="row1">
    <div role="gridcell">Cell</div>
    <div role="gridcell">Cell</div>
  </div>
</siv>

<script>

function axElementById(id) {
    return accessibilityController.accessibleElementById(id);
}

test(function(t) {
    const axPresentation = axElementById('row1');
    assert_equals(axPresentation, undefined);

    // Change role attribute
    document.getElementById('row1').setAttribute('role', 'row');

    const axRow = axElementById('row1');
    assert_equals(axRow.role, 'AXRole: AXRow');


}, "Role presentation changes to role row.");

</script>
